+++ Changelog.md
@@ -1,6 +1,8 @@
 ## 0.0.44
 + Added `target` argument in `@compile` decorator for specifying specific targets to compile for.
-+ Changed `fxn.predictions.stream` method to be a plain `Iterator` instead of `AsyncIterator`.
++ Refactored `fxn.predictions.stream` method to return an `Iterator` instead of `AsyncIterator`.
++ Refactored `Acceleration` enumeration to string literal type.
++ Refactored `RemoteAcceleration` enumeration to string literal type.
 
 ## 0.0.43
 + Added `fxn compile` CLI function for access to the Function compiler proof of concept.